requirements-rag.txt
# RAG System Requirements
# These are in addition to the main requirements.txt

# Core RAG dependencies
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
tiktoken>=0.5.0

# Backend API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# AWS integration
boto3>=1.28.0
mangum>=0.17.0  # For Lambda deployment

# Local LLM (optional for development)
httpx>=0.24.0  # For Ollama API calls

# Data processing
PyYAML>=6.0